% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbSafeNames}
\alias{dbSafeNames}
\title{Change R variable names to db-safe versions.}
\usage{
dbSafeNames(names)
}
\arguments{
\item{names}{The vector of variable names}
}
\description{
Change R variable names to db-safe versions.
}
