% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_text}
\alias{get_text}
\title{Get text from a file and clean for searching}
\usage{
get_text(file)
}
\arguments{
\item{file}{Path to the file to be read}
}
\value{
A single-element text vector of clean text for searching
}
\description{
Get text from a file and clean for searching
}
