% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gn_filter_natural_places}
\alias{gn_filter_natural_places}
\title{Keep 'natural' feature types in df}
\usage{
gn_filter_natural_places(df)
}
\arguments{
\item{df}{A data frame from one of the \code{us.geonames} search functions}
}
\description{
Keep 'natural' feature types in df
}
