% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gn_read_txt}
\alias{gn_read_txt}
\title{Read a text file and clean up spaces}
\usage{
gn_read_txt(f)
}
\arguments{
\item{f}{Path to the text file to process}
}
\description{
Read a text file and clean up spaces
}
