% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gn_search_all}
\alias{gn_search_all}
\title{Search a file f for all geonames using \link{fastmatch}}
\usage{
gn_search_all(text, ngram_min = 1, ngram_max = 7)
}
\arguments{
\item{text}{Text to be searched for geonames; should be 'clean,' i.e., no
eol characters, multi-spaces replaced with singles, etc.}

\item{ngram_min}{Min ngram length to create of each \code{texts} (default = 1)}

\item{ngram_max}{Max ngram length to create of each \code{texts} (default = 7)}
}
\description{
Fast reverse-lookup search of all 2.27M place names in \code{geonames}.
Whereas \link{gn_search_all} searches the 2.27M geonames against the
\code{text}, this function tokenizes
}
