% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gn_search_all_stri}
\alias{gn_search_all_stri}
\title{Search a file f for all geonames using stringi}
\usage{
gn_search_all_stri(text)
}
\arguments{
\item{text}{Text to be searched for geonames; should be 'clean,' i.e., no
eol characters, multi-spaces replaced with singles, etc.}
}
\description{
Costly but thorough search of all 2.27M place names in \code{geonames}.
}
