% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gn_search_state}
\alias{gn_search_state}
\title{Search text for all geonames from a given state}
\usage{
gn_search_state(text, states)
}
\arguments{
\item{text}{Text to be searched for geonames; should be 'clean,' i.e., no
eol characters, multi-spaces replaced with singles, etc.}

\item{state}{Two-letter state abbreviation, e.g., VA for Virginia}
}
\description{
Why search all of geonames if you know the state(s) of interest?
}
