% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gn_tokenizer}
\alias{gn_tokenizer}
\title{Tokenize word ngrams (1-7 default) of text}
\usage{
gn_tokenizer(text, min = 1, max = 7)
}
\arguments{
\item{text}{A character vector to be tokenized}

\item{min}{The minimum ngram length to create (default = 1)}

\item{max}{The maximum ngram length to create (default = 7)}
}
\description{
The default of unigram to 7-gram tokenization was chosen because 99% of all
placenames in the geonames gazetteer are <= 7 words long.
}
