% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gndb_search_state}
\alias{gndb_search_state}
\title{Search text for all geonames from a given state (db version)}
\usage{
gndb_search_state(con, text, states)
}
\arguments{
\item{con}{A db connection from \link{DBI} or a \link[dplyr]{tbl} object; the
tbl is preferred}

\item{text}{Text to be searched for geonames; should be 'clean,' i.e., no
eol characters, multi-spaces replaced with singles, etc.}

\item{state}{Two-letter state abbreviation, e.g., VA for Virginia}
}
\description{
Why search all of geonames if you know the state(s) of interest? This uses
a database connection rather than relying on reading the entire 2.3M rows
of geonames from file.
}
