% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{gn_leaflet_basic}
\alias{gn_leaflet_basic}
\title{Basic leaflet map of points identified in a \code{us.geonames} search}
\usage{
gn_leaflet_basic(df)
}
\description{
Basic leaflet map of points identified in a \code{us.geonames} search
}
